requirements.txt
fastapi
uvicorn
sqlalchemy
yfinance
aiohttp
beautifulsoup4
Pillow
groq
python-multipart
feedparser
apscheduler
httpx
jinja2
pydantic
curl_cffi
